fix: address P1 review findings (batch 2)
- Use float division in DetectPosture to fix boundary rounding (spgr-uqx.38)
- Report all unknown stages in ValidateTransition error (spgr-uqx.47)
- Surface skipped files in scanner Tier1/Tier2 results (spgr-uqx.11)

+++ internal/authoring/posture.go
@@ -25,12 +25,12 @@ func DetectPosture(messages []string) specv1.Posture {
 	for _, m := range messages {
 		total += len(m)
 	}
-	avg := total / len(messages)
+	avg := float64(total) / float64(len(messages))
 
-	if avg < driveThreshold {
+	if avg < float64(driveThreshold) {
 		return specv1.Posture_POSTURE_DRIVE
 	}
-	if avg > supportThreshold {
+	if avg > float64(supportThreshold) {
 		return specv1.Posture_POSTURE_SUPPORT
 	}
 	return specv1.Posture_POSTURE_PARTNER
